+++ prjiSpanFinal/ViewComponents/CalTotalPriceViewComponent.cs
@@ -1,0 +1,16 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace prjiSpanFinal.ViewComponents
+{
+    public class CalTotalPriceViewComponent : ViewComponent
+    {
+        public async Task<IViewComponentResult> InvokeAsync(int? id)
+        {
+            return View(id);
+        }
+    }
+}
