+++ prjiSpanFinal/Views/Delivery/_DeliveryCheckoutPartial.cshtml
@@ -58,7 +58,7 @@
         <div class="fw-bold" id="finalShip">郵寄寄送</div>
         <div class="text-black-50" id="finalAddress">106台北市大安區復興南路一段390號2樓</div>
         <div class="d-flex justify-content-between">
-            <div>運費:<span class="text-danger">$0</span></div>
+            <div>運費:<span class="text-danger">$60</span></div>
             <button class="changeShip border-0 bg-transparent fs-6 text-primary">變更</button>
         </div>
     </div>
